Move 'pipetask run-qbb' implementation directly into script.
This completes the removal of CmdLineFwk.

It also drops the test for _make_execution_resources, since that's now
just duplicated across the 'run' and 'run-qbb' scripts: it's one line,
and the static typing we've got now is at least as good as that
type-focused test.
